refactor: SearchDialog에서 filter component 분리
